Updated the TColour class with some additional operators and removed teh Limit() call from the others.  Also exposed the Limit() routine as public and added a constructor for a float4 type.
Added support for the import of the TColour type which was missing from TLFile::ImportBinaryData().  Also fixed an issue with the number of elements for the g_DataTypeRefCache array being incorrect meaning the last few types would not be seen as valid types ffor import.

Added support to the TFileTimeline to import TColour24, TColour32 and TColour64 and export as a TColour type to the asset.

Added colour support to the TTimeline class so that colours can now be interped.

Updated the TRenderNode to process a colour message and set the colour appropriately.  
NOTE: 
This will not affect meshes that have colour exported.  
Also currently the message is part of the SetTransform (incorrectly) and is currently like this because the commands from the timeline have child data which in the case of the colour want to be commands in their own right.  I suspect most properties will be the data elements like the elements of the SetTransform but that works slightly differently in that it packages all three into one message.  Anyway, there are two options here a) have a SetProperty message which than uses the data as sub-message elements for setting the properties or b) intercept the data and use them as the message directly.  The former is the easiest option but involves having extra checks in the receivers ProcessMessage routines and the latter is how most of our messaging has evolved to work but will involve handling special cases within the timeline system.  I am probably going to update the timeline system to use method (b) so expect further timeline changes in this area soon.