Added multiple shape support to physics nodes. Outside the physics node's these are known as "collision shapes" and generally just "shapes" inside the physics node code. In Box2D they are "fixtures". Each collision shape has a shaperef now, this way you can add and remove separate/specific collision shapes from a node. This ref is also used in CollisionInfo now so you can tell when a specific shape receives collision (and with what shape of another node). This allows specific code for a collision with a sensor on an object that also has collision shapes. If you want to assign a ref to a collision shape from data/message add a "Ref" child data with the ref. If you want the shape to be a sensor then add a "Sensor" bool data to the ColShape data along with the ref. The Sensor flag is no longer on the physics node, and is set individually on shapes. If you're creating a shape from a datum ("ColDatum" on a scene node object) then the child data's are copied (ie. sensor). The ref is automaticly set to the datum's ref.
Note (not previously mentioned in log) if NodeA collides with a sensor on NodeB, NodeA gets no collision message. (as there was no reaction)

Added a TLCore::IntegerToPointer conversion function to go with PointerToInteger. I'm using this to convert a void* pointer to a u32. Not for storage! runtime only! Pointers are not cross platform!

Added some functions to PtrArray to do simple TPtr comparisons (searches).

Removed the code for TPhysicsNodeSphere, all this functionality in 2D is part of box. If we need this in 3D again, we can restore this code, if we don't have some 3D physics engine which does that instead.