Added "bool" binary type macro. Added Integer data importing, supporting u8,s8,u16,s16,u32,s32. Fails if the number extracted from the string is over their limits. (Fails and breaks in debug only, release will blindly cast the string to it's type)
Added <Bool> data importing, goes off of the first character in the string. 1,t,T for true. 0,f,F for false. So valid examples are False,TrUe and Fish. 

Removed the type array and replaced with a switch statement to say what types are supported. (Much better)
Looking at the code this function (TLFile::GetDataTypeFromString) is redundant as it's not used without TLFile::ImportBinaryData, we could merge these together... left for now though.

In the functions TBinaryTree::CopyDataTree and ReferenceDataTree I've removed the line which empties the existing children. (The root data is still overwritten, I may change this to a merge/write too) Looking at the code this won't make any difference to existing stuff, but now we can merge two TBinaryTree's together into another without losing any data.

Added a TMessage (Just realised I should probably change this to a TBinaryTree, not a message) to the Sceheme Import/ReImport (should get rid of Reimport too btw) so you can add "common" data which is inserted into the initialise message of each node imported to the graph.
I've added this specificly so I can add a RenderNodeParent (RNParent) field to every scene node in a scheme to ensure the render node for that scene node is created in a visible part of the tree. Not neccessarily the best way of doing things... but this needs thinking about along with trac bug #57
Off the top of my head I'm not sure what other use this could be... but I needed it.

Fixed XML parser to not throw up errors/fail parsing when it finds certain characters in <!-- comments -->. Any <> angle brackets inside comments will probably break still though.