Moved <data> XML parsing into it's own function in TLFile. Added support for a whole data tree (surprised we've not used this yet), ie. <data><data>
Added support for <data> in a scheme file, this data (in the root of <Scheme>) just goes into the asset's unknown data (TAsset::GetData())

Added debug break check for when we write some data, but the hint isn't set to catch problems with the hint system.

Fixed hint being invalid on TBinaryData when writing a string.

Changed the marker on physics nodes to say the body transform is out of date to a bitmask. It doesn't reduce the amount of things we change (not a performance hit) but catches when scale changes on the physics node, which requires a rebuild of all the collision shapes (This hasn't been implemented yet as it's not required, but I've added this check so I don't forget itin the future)