Removed a couple of redundant includes from the TApplication.cpp file.  I think the main reason this rebuilds often is that it includes the screenmanager which includes a load of other stuff so I'll look into reducing this a bit.
Added a few new static trefs to the TLCore namespace.

Added a generic SendMessage(TRef Recipient, TRef Manager, TMessage Message) routine to the core manager.  This means you can now send messages through the core manager to graph nodes or to managers themselves eventually or some other child of a manager as required. This is useful to avoid duplicating code to test manager ref's for sending messages to nodes in any of the graphs where you only know the graph ref for example, or where you use a manager ref in generic code and adding includes to be able to use the global manager pointers for all possible managers would not be suitable, especially in the engine code.  To compliment this there is now also a generic SendMessage(TRef Recipient, TMessage Message) on the TManager class which is specialised on the graphs as a wrapper call to SendMessageToNode(TRef, TMessage).
NOTE: For node creation and shutdown you will still need to test for each graph required but I intend to add something to cater for these as well at some point.

SetProperty message

To correct a problem with nodes being created multiple times via timelines when they shouldn't be - usually after calling the init first which creates a node on first frame for example and then again on first update - the timelines now keep track of the node refs they create for each of the graphs.  The timeline instances also  subscribe to the graphs (if they create any nodes) to be informed when nodes are shutdown that may be the ones they created so that they can keep their list in sync.  Although this corrects the problem, it also puts a few limitations on node creation via timelines in that you cannot now just loop a timeline and create numerous nodes like an emitter using the same XML data because the second time a create is attempted where the node ID has already been created the create command will be suppressed. Depending on how we eventually use this system will dictate whether this needs reverting back and a different solution finding for duplicate node creation.

Implemented the SetProperty(TMessage Message) routine.  This eases a few things:
* Where we needed to put property data under one Message ref to mirror how the timeline XML is structured
* Preference to avoid using the Initialise message for setting properties when it is designed to be a one-off setup routine.
* Avoid overloading ProcessMessage for lots of nodes that handles a lot of messages already
* Keep property setting contained within a defined routine for ease of checking in the code

The SetProperty has been added at the TGraphNode level and is called via the TGraphNode's ProcessMessage() rotuine which handles processing of the Initialise and Shutdown messages.  Essentially for simple property setting, which may include additional processing such as invalidating bounds or sending messages for property changes, you should now overload and use this routine ensuring you also call the super or parent SetProperty() routine as well.  I have set up an example in the TRenderNode which sets the rendernodes colour property.  You can also call this routine directly from routines you currently use to set properties like the Initialise() and pass the initialise message as within the SetProperty you simply test for importing the properties your node will be interested in.  This will only real work if you know the message contains properties otherwise it will not do a lot other than slow things down.

NOTE: Currently the transform properties are separate and handled via the SetTransform message.  There is no reason we cannot put the transforms under the same SetProperty message which I will look into next week.  I believe only the timelines use this message anyway.

Also Graham you may want to add a TGraphNodeBase SetProperty for adding properties that don't get read... perhaps?  I'm not totally sure how this works from the Initialise message so I'll leave that with you.


GetProperty message

To compliment setting properties via the messaging system I have added preliminary work on the ability to get properties via the messaging system as well.  This will eventually be useful for things like editors where you would like to query the properties of nodes where you might not have access to the node(s) in question or having to cast a node pointer to the appropriate type for use would be unsuitable.  I have therefore setup the ability for graph nodes to process a GetProperty message and call a GetProperty(TMessage Query, TMessage Response) routine.  

There are two options for how this will eventually work:  

* Send a query message to the node with all of the data you want to know about, i.e. colour or translation, and then the node adds the required information to a response message which is automatically returned as part of the TGraphNode processing for the GetProperty.
* The node simply adds all of it's properties to the response and the sender is then responsible for parsing the response message to filter out the info it wants to know about.

Both have benefits and limitations.  The former would allow narrowing down particular information and limit the data being sent around but would also mean needing some other system for obtaining all properties without having to have the initial message queried and also you might not know what properties are available on a specific node.  The latter means the node doesn't have to do much other than add all of it's properties to the response message but means a lot of potentially redundant data could be sent around, but does mean the initial sender will have all the information it needs to do whatever it needs to do with it. 

Because the GetProperty is not likely to be used just yet, we can have a think about which way would be most useful but I would be inclined to use the latter option as it is likely be the most flexible.