Added StoreTimestamp() and RetrieveTimestamp() routines to the core manager to store/retrieve timestamps within the core manager's hardware information.  This will be used for accessing generic application time stamps such as the application start time, for statistical purposes and optimisation guidance.  Game specific times could also be added to the core manager but it would be preferable to use the TGame or TApplication class for this data.
Added storage of the following timestamps to the core manager:
* TSStartTime - The time the application is started.
* TSInitTime - The time the initialise finishes, before any update starts.
* TSEndTime - The time the application begins the shutdown.

These can be used to determine length of time for initialisation and gameplay as required.

Fixed an issue with the startup time being setup twice - once from the time manager and once from a specific call to the TLTime::Platform::Init() during initialisation.  The specific call has been removed.

Added a RandomNumberManager to initialise the random seed correctly post the timemanager initialisation now that the specific time init call has been removed.