Put TManager class into the TLCore namespace. (Lots of file changes) Changed the graph managers to initialise the manager super class with specific refs (NOT dictated game-side). To use this I've added a CreateAndRegisterManager that doesn't take a ref for these classes.
Added Debug_CheckNotOnlyPtr() to TPtr classes. Use this to check TPtr<X>& params. If we have this for a param, and only 1 in the counter, then a variable has been passed in and a TPtr has been created. This will crash/corrupt memory when the func returns as it will delete the contents. The function doesn't return anything, intended to only use during debug testing to catch errors (usually code errors and not runtime stuff)

Moved lots of graph node and graph functions into the base classes so they can be modified more easily outside the template code (be nice to reduce code in that header into source files where appropriate) and makes things a little simpler to edit. Includes moving quite plain simple stuff (eg. message, manager handling) to the base class, and the contents of TLGraph.h become much more graph-functional.

There is a small hacky function on the base node class called GetRenderNodeRef(). I'm currently using this in the scheme editor until I can decide the best way of getting what render node to use for widgets for the generic scheme editor. I could use the node data... but that won't be setup by default in most cases, and is a little overcomplicated. The rendernode ref should be returned on a class level (ie. virtual function) and not at an instance level (data)...
Doesn't work for physics nodes yet, but in a bit I will make it find the owner scene node, then get it's render node... bit over the top, but called rarely (at least it is for the scheme editor).

Moved the old code from the node base class initialise into SetProperty, which stores unread data onto the node.

Changed the language refs to be static and used them where code WASNT using the predefined refs...

Added an AddTransform function to the TTransform class which modifies values of a transform by their own values (imagine in world space). Using Transform() with a transform makes changes in a local space... to go with this DoTransform message for scene and render nodes (and one day physics nodes) now does the AddTransform. A LocTransform (LOCal Transform) message is now handled which transforms the current transform locally... so if you want the object to move right relative to it's heading, it will.

Corrected the SceneNode Object code so that OnTransform will always relay changes onto the physics node, except in special circumstances. This means the reliance on SetTransform is gone. Most of the SetTransXXX code has gone now as it was redundant and the code on the Transform scene node was more than sufficient.

Corrected duanes "colour" example of the use of SetProperty in the render node. I've also moved the property-changing functionality into the SetProperty function. (Including an example of changing the transform... a replacement of SetTransform message).
The Initialise of the RenderNode just calls SetProperty... I don't know if we intend to move this to a lower level (ie. base graph node) where the Initialise just calls SetProperty with the init message. 
Every node WILL NEED to do this for scheme/timeline/xml/etc property initialisation...

Added a new TLGame::TSchemeEditor class to the game lib. Currently can only modify one graph at a time, but it can do any graph.
Adds a widget to every node in the tree you specify, that widget can then catch you dragging the node around. It sends an "EdtStart" message to the node when first selected (eg. to disable physics so you don't compete with gravity) and then an "EdtEnd" when let go (to restore state)
Next step is to add a custom gui and message handling, and a node instancer (via a UI)

This class will be able to be bolted onto any game. I shall add the functionality to the physics test game (or possibly the entire test app...) when I have implented GetRenderNode on physics nodes. 

Removed the graph refs in the timeline code, and replaced it with a generic function to fetch the appropriate graph. I did this to remove those refs rather than meddle with your code duane. :)