Removed the old TLGRAPH_OWN_CHILDREN macro and changed graph to always own children (ie. what we've been using for 6 months)
Renamed TInputInterface to TWidget.

Added a TBinaryTree constructor to TWidget. This is because I wanted to add more [optional] parameters and the constructor was getting to a silly length. The old one is there for simple widget construction, but we can add more things and more flexibility etc with a TBinaryTree version.
The Initialise code has been removed from the widget as it's now essentially just a subscription. This also means setting up variables from the TBinaryTree in a widget-derivative won't be a pain (ie. if we needed to do it before Initialisation before)

Changed the secondary click check (first is always a fast sphere bounds test) to use a specified datum. This defaults to the bounds box like we tested before, but it means we can use datums on rendernodes for widget clicks now, which simplifies buttons and UI. We can have a whole gui/menu designed in one mesh (one .svg) and have multiple buttons on it which would be visually setup in inkscape etc.
These datums are pre-fetched before processing a queue of clicks which is a good speed up. (and apllicable to future changes to the widget processing)

I've simplified the tests to see if a TClick is up/down/move by adding an ActionType which is a ref which is either Down,Up,Move. This is set when the click is queued up and heavily reduces the use of the TWidgetManager::IsClickActionRef/IsMoveActionRef and means we don't have differing versions of detecting if a mouse click is up or down (==0.f, <NearZero etc)

Fixed an issue where a widget would queuue up tons of clicks/moves when the render target it was in was disabled.
This is more complicated than just ditching the queue though; if the render target is disabled (will expand to if rendernode disabled) mouse moves and down's are ignored. A mouse UP is still processed IF it was the previously processed action, otherwise ignored. This makes widget-handlers able to "clean up" when the mouse is up rather than just thinking it's stuck down just because the rneder target is gone.

One required expansion on this is to catch when a render target is disabled so a fake "up" click can be created to instantly stop the mouse click. Currently if the RT is disabled and you still have you finger down, the widget thinks its still down, but it should be gone as the widget is gone.

Added line intersection virtual function for shapes. (a line is not a TShape... yet, not sure if it will be)