Added support for an interface for the scheme editor; To enable this you need to supply a scheme which contains the "gui". This then creates a render target and places your gui in it. The scheme also contains various bits of data;
Widgets can be created with multiple <Data DataRef="Widget">. The widget contains data that is fed straight into the new TWidget binary data handler (render node, datum, action to send on click etc). Currently widgets are always buttons (until a widget factory is done... and is required)
The action in the widget sent out will correspond to editor commands (eg. "clear", "close", "undo") so you can add/remove features in your gui via the scheme.
Close is currently the only handled command.

"NewNode" data's will also be supported which will dictate what types of nodes can be created, will contain extra data (eg. friendly strings, icon textures/meshes). This hasn't been implemented yet and I haven't decided if the new-node draggable interface will be constructed by overloading the scheme editor, or a generic solution which can be customised with different kinds of nodes, specific overloaded icon creation etc.

Added TBinaryTree construction to TWidgetButton.

Added a KeepShape option to TShape's transform function. This is mainly to keep box shapes as boxes (though might turn out to be useful for other things), so that in that case a simple box isn't converted to a Polygon which will require a more complex and expensive point/line intersection tests (and will need writing/copying from box2d)

This KeepShape option is also on TRenderNode::GetWorldDatum. This is because most of the time when we fetch the bounds box datum we want it as a box (for efficiency of widget usage)