Made some little changes to duane's changes in the scrollable render node;
Changed the scroll float2 to a TTransform in TRenderNodeScrollableView, this saves creating extra transforms etc on the stack. The original float2 part of the translate is still exposed for easy access/modification

Added a BeginScene and EndScene so you don't need to do lots of extra calculations/CPU work when "untransforming" the changes by the scroll view, just restore the state of the scene :)
I've removed the temporary scene transform param from PostDrawChildren as I assumed it's never going to be very useful as the transform won't have changed since it PreDrawChildren.
Added the render target rendering this node to those functions to do the Begin/EndScene.

In the render target DrawNode I've changed it so it doesn't do any of this stuff unless it actually has children. This saves creating another TTransform on the stack, redundantly calling these functions on nodes that don't support it, and means you don't have to check for children in the functions themselves.

Added some comments to the on-scroll message thing, the children will need to be invalidated but it's a little complicated. These issues will show up when we want to use the children as widgets...

(sorry to intefere with your code duane, but it's simple improvements, let me know if any changes don't quite make sense)