Added "icons" to the scheme editor to allow the creation of nodes into the graph. The actual icon render nodes are a bit of a hack at the moment until I come up with a good way of specifying them in a custom way. Currently they're put in at a generic location, but psuedo dragging functionality will go in next.
Added TBinary construction to TWidgetDrag.

Split the alignment transform and normal transform of text render nodes. This simplifies a whole bunch of things and means we can now align around a point (ie. without a box) and we can offset from that box (eg. for faked multiple lines). When the transform changes, the class jumps in and applies the alignment too. This *might* not quite work with the DoTransform functionality, in which case the text render nodes will catch that message and handle it manually.