Changed the GetUser function to return a reference to a TPtr. Got carried away and cleaned up/simplified some code, renamed some variables and functions etc... spotted some bad things...
Added a simple HasChild function to TBinaryTree to see if a child already exists.

Moved the auto-add unread data which gets added to the node data, data from InitMessages out of the base functions and into specifically used functionality in the non-base graph.
Changed the setup of the use of SetProperty and Initialise. Now, when a graphnode gets an Initialise message it calls SetProperty() to "initialise" variables on the node. Then the [soon to be] one-off function call to Initialise() is made. After that any data that hasn't been read in the Initialise message is stored on the NodeData. 

For the SetProperty message handling, and unread data in there is also added to the node data. This code for the Initialise and SetProperty that stores node data is NOT in the base NODE functions so you don't have to worry about call orders causing redundant data being written to the node data.

Added mouse tracking to the scheme editor. This allows me to catch a drag from an icon and fake a widget-like drag of the new graph node. This tracks the mouse until the mouse button is let go and the node is "unselected", this is mostly for show as the node is already in the scheme root at this point. Should work with any mouse button and keep that button for the mouse tracking but untested.

Changed the base widget SendActionMessage function to just pass in the related TClick and what action to send. More data is sent through with the action now including the original input action.

Moved some code from Initialise of Physics node and scene nodes to SetProperty.

Removed the now redundant call to SetProperty in the render node and scene node's initialise.

Added some code so if we try and disable/enable a physics node that isn't in the graph yet a message is sent. Possibly we always want to send a message... 
Fixed it so you can disable the physics node on a scene node at init (in SetProperty) and it won't be re-enabled automatically.