Fixed asset export so if it's a 2nd export the data is cleared (ie. the file asset data is fresh for a new asset export). This was causing a bug when you re-saved an existing asset.
Added a check to ReferenceDataTree to ensure we're not overwriting existing root data (or that there is no clash).

Changed graph ExportScheme to export to an existing asset instead of creating a new one. (You can do that yourself). Removed the GetNodeData() function that had a param to update the node data. Instead there is a seperate UpdateNodeData() function. This is called when a node needs to be serialised. Any data you have that isn't stored (or maybe out of date) in the node data needs to be updated here.
One example (needed it anyway) is the scene node transform, UpdateNodeData here removes old transform data and writes the latest transform.

Changed the file-asset data root's data ref (The TBinaryTree that an asset exports to, then is serialised into a plain file. Before it was set to the file ref, I spotted a ref in a data tree as "   AD" and couldn't work out what it was. The file refs are temporary (can change for different app boots) we probably shouldn't save this. As this data ref is saved to files (it's never actually used for anything) I've renamed the ref to "Tooot", so if you ever spot that, you know it's come out of an exported asset file. Shouldn't cause any problems, but makes more sense if you see that ref.

Scheme editor handles "save" command now, and relays it out. I MIGHT put the saving functionality into the scheme editor, but for now the client has to do it in case there's extra stuff to do. 

Extra data can now be attached to widget action messages. Example of this is that the WidgetDrag now sends out a bool of "DidDrag" with the actions. If this is false, the user essentially clicked the widget, and didn't drag it. This may be useful later on for having widget do multiple things depending on what the user does. (eg. menu on click, drag on drag)