Added GetScreenPosFromWorldPos routines to the TScreen class.  These are the opposite of the GetWorldPosFromScreenPos routines and will (eventually) handle the different rotations of the screen when converting the final screen position.
Removed the bFlipY parameter from various routines as it was causing problems later down the line with the rotated screens and is now redundant as the flipping is done in the new GetScreenPosFromWorldPos where it is more appropriate.

FYI The scrollable render node still works in portrait mode but I am still working on the landscape mode which is proving a bit more difficult to correct.