Revised the application/game flow and state stuff. I may want to implement more changes as the application defaults aren't flexible enough really and some of it I feel is erroneous (Eg. enter/exit game - means you can't use in-game transitions as the game is lost)
I KNOW this will break other games, but have not yet fixed them, though the only real change I think will be the CreateGameObject which just returns a new TGame object now and does no initialisation and setup.

The game is now properly registered (and now unregistered) as a manager so it can be subscribe to/from on nodes.

The GetManager function now returns c-pointers not TPtr's. To get a TPtr call GetManagerPtr.

More flow changes to come I'm sure as I add more flow to rocket game properly (the seaman flow was all in the game and I'm trying to remove that)

Added a "PubTo" and "SubTo" message handling for nodes. Inside you add "Manager" data's to subscribe to/from a manager or graph. And "Node" to subscribe to another node. An additional "Manager" data needs to be below the node to work out which node in which graph it is you want to subscribe to/from. 
This removes the need for specific "NodeAdded" handling from an object that wants to subscribe to a node it has created.

Added SetMode() to the TApplication to tell it to change to a new mode on next update. Added a siilar SetMode() to the TGame to do the same thing.
I feel SetGameMode() function will need to be renamed to be something more specific.(This is used to set a ref which dictates what kind of TGame to create. used in the test app)