Implemented base TSceneNode_Object life handling.  By default all objects derived from the TSceneNode_Object have a life value of 100.0f and can be 'killed'.  The following virtual routines can be used to handle any extra life change requirements:
* Bool CanChangeLife(const float& fLifeChange) - called before altering the life of an object.  Return False to prevent any changes to the objects life making the object indestructible.

* void OnLifeChange(const float& fLifeChange) - called after life has changed with the amount of life change.  Use this to show damage/healing via graphics or sounds.  By default this routine checks for and trigger the OnDeath() so be sure to call the super routine if you overload this.

* void OnDeath() - called when the life change drops the life to 0.0f or below. By default this will remove the scene node.

There are also accessors on the TSceneNode_Object to get and set the life.  You can also add a 'Life' property to an Initialise messages which will set the life as required within the SetProperty() call.

If you want to change an external nodes life at runtime send a 'LifeChange' message with a float for the amount of life to change by.  This should be done when doing damage to another object or healing another object for example.