Found an alternative debug break for the ipod in the forums. Not implemented but code & comments are in.
Added the ability to specify geometry as "AndDatum" instead of "Datum". This creates geometry AND a datum so you can use graphics directly as datums now.

Added some extra debug and a small speedup to the timeline code.

Fixed some stack-corrupting code in the polygon shape when it had more than 8 points.

Added an OnRenderNodeCrated() function to the scenenode object which gets called when the object has a valid Ref for the render node (though not yet in the graph)

Added missing pragma once to TSceneNode_Timeline.h