Renamed some functions and variables/params for some of Duanes screen/world/view functions. The "spaces" of the units were slightly out. This is just for a bit more clarity between World,Ortho,RenderTarget,View and screen space (in that order :)
Added a function for a simple conversion of world unit SIZE to pixel SIZE. This should NOT be used to positional information, just scalar.
Corrected the world point size -> pixel size conversion for point sprites. Also skips over the texture environment changes when using point sprites and not using textures.

Exposed the GetAllocSize function for the FixedArray, doesn't need to be protected.