Added support for box2d's edge chain shapes. These are slightly more complicated than normal shapes (it's not even a shape). You can create a edge chain with a TPolygon2D which has more than 8 points (box2d's polygon limit). I may expand this to support the new line shape and non-convex polygon shapes. More progress will follow when I find out what the main restrictions of this shape are and if it behaves differently. It may have to only be a static shape...
Added a TShapeLine. This is mostly just to have a simple routine to pull the edge shapes from box2d.
This is a single line, I may expand this to a linestrip "shape" or remove it entirely.
Added various bits of support for the Line shape.

Moved the physics collision shape code into the old TCollisionShape class.