Removed the Bool parameter from the TBinrayTree ReferenceDataTree() routine.
Implemented the UpdateNodeData for both the TRenderNode and TPhysicsNode classes.  This will export various properties to the node data binary tree.  There may be variables that are still needed for export which can be added as required.

Updated the TSceneNode_Object UpdateNodeData() routine to update both the render and physics nodes using the same routine and then reference that data under two new subtrees "RNode" and "PNode".  This means the render and physics node data will now be reflected in a scheme asset file.

The SetProperties of the TSceneNode_Object now imports a "RNode" subtree for the render node data and "PNode" for the physics node data.  If these don't exist it will work as before checking for the type and parent within the message as required for the CreateRenderNode() and CreatePhysicsNode().

Added a SetData(STORAGETYPE& flags) routine to the TFlags class for raw setting of the flags rather than having to use indices.

The render and physics nodes can now import their flags as a single raw value (using the new SetData() routine) with a ref of "RFlags" and "PFlags" respectively in the SetProperties call.  If these items fail to import it will revert back to importing the flags on a per flag basis using "RSet", "RClear", "PSet" and "PClear" as required.