Changed the ImportScheme code to use a BinaryTree for the extra data instead of a TMessage
Added support for spawing a scheme from an "icon" in the scheme editor. The scheme is put underneath an empty scene node (which has to be an object...) and instanced with non-strict node refs.
There seems to be slight issues with it atm, it's spawning in the wrong position (not dragging with the mouse) and once it is in, selecting/dragging the widget is a little off...

Instead of using a "Type" in the scheme-editor scheme XML use "Scheme" with the ref of the scheme you want to spawn.

Added "Child" support in the render node data. When UpdateNodeData is called it exports all children's data into an extra "child" tree. In the scene node this goes underneath the "RNode" data. The entire render tree is then re-created when the scene node is re-loaded.
Will have to think about this implementation a little more when we export other graphs as well as the scene graph. (Maybe just not export/save render nodes which have an owner...)

Added an Initialise() function to timeline instances that doesn't require an init message.

Added a TRenderNodeTimeline node. This is a simple node with a timeline attached to it. Now you can instance an animated render node which requires no manual code. The data "Timeline" tells it what timeline to attach. A "Play" property exists to stop/start animation. A "Reset" message will put the timeline time back to zero