Big asset change; Assets can now have the same AssetRef as long as they have different types. This means we can have a "stars" particle AND a "stars" texture co-exist in the asset system. The external accessors have changed now.
A part of the LoadTask and file system code needs changing still before we can actually have multiple asset-source files with the same filename though, so don't start duplicating asset names just yet.

GetAsset and GetAssetPtr are now the primary functions to get an asset. The default params (SyncTrue) will block load the asset. SyncWait will start an async load but will NOT return an asset unless it's already loaded. SyncFalse will NOT try to load the asset, so will only return one if it's loaded.
There are various overloaded versions of the functions to return more native pointers and can deduce arguments. The most common accessor (which should be used in game code) is TLAsset::GetAsset<TLAsset::TMesh>("mymesh");
This will replace any previous LoadAsset("mymesh", TRUE).GetObject<>() code.

TLAsset::LoadAsset still exists, but returns a SyncBool as to the file's loading status. It has a param to allow block loading.

To aid this, a function has been added to all asset types; GetAssetType_Static. Probably not much use outside of the templated functions though.

I've fixed some minor issues with the audio asset importing, previously whilst the asset was being imported from an asset file (part of the load) it would then call a function which would invoke the load of the same asset (whilst it's loading) this gets the system a bit confused and means the audio CreateBuffer function would be using an asset which isn't technically considered loaded.

I've renamed the TAssetTimeline asset to TTimeline (file rename too). I've made some changes to the TTimelineInstance class too which now stores a pointer to the asset at construction/set instead of looking it up all the time.

I've merged the ArePreloadFilesLoaded() and PreloadFiles() in the app boot state to the same function to reduce code duplication. 
Warning: the GetPreloadFiles function on the TApp has changed param, so this needs updating on apps.

I've made the SchemeManager manager ref fixed to "SchMan" so that it is now consistent over projects.

Fixed the compile error in the local file sys for the ipod.